requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.5
django-magiclink==1.2.0
django-phonenumber-field==6.1.0
idna==3.3
install==1.3.5
jwt==1.3.1
packaging==20.9
phonenumbers==8.12.48
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
